+++ WorldOfTheThreeKingdomsEditor/NewEventWindow.xaml.cs
@@ -1,0 +1,28 @@
+﻿using GameObjects;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace WorldOfTheThreeKingdomsEditor
+{
+    /// <summary>
+    /// Window1.xaml 的互動邏輯
+    /// </summary>
+    public partial class NewEventWindow : Window
+    {
+        public NewEventWindow(bool editing, DataGrid dataGrid, GameScenario scen)
+        {
+            InitializeComponent();
+        }
+    }
+}
